+++ src/test/java/com/bishnet/cucumber/parallel/runtime/CucumberRuntimeFactoryTest.java
@@ -19,7 +19,7 @@ public class CucumberRuntimeFactoryTest {
 		featurePaths.add("a/relative/path");
 		List<String> cucumberPassthroughArguments = new ArrayList<String>();
 		RuntimeConfiguration runtimeConfiguration = new RuntimeConfiguration(0, cucumberPassthroughArguments, null, featurePaths,
-				null, false, null, false, null, false);
+				null, false, null, false, null, false, false);
 		CucumberRuntimeFactory runtimeFactory = new CucumberRuntimeFactory(runtimeConfiguration, Thread.currentThread()
 				.getContextClassLoader());
 		runtimeFactory.getRuntime(cucumberArguments);
@@ -33,7 +33,7 @@ public class CucumberRuntimeFactoryTest {
 		featurePaths.add("classpath:");
 		List<String> cucumberPassthroughArguments = new ArrayList<String>();
 		RuntimeConfiguration runtimeConfiguration = new RuntimeConfiguration(0, cucumberPassthroughArguments, null, featurePaths,
-				null, false, null, false, null, false);
+				null, false, null, false, null, false, false);
 		CucumberRuntimeFactory runtimeFactory = new CucumberRuntimeFactory(runtimeConfiguration, Thread.currentThread()
 				.getContextClassLoader());
 		runtimeFactory.getRuntime(cucumberArguments);
@@ -45,7 +45,7 @@ public class CucumberRuntimeFactoryTest {
 		List<String> cucumberPassthroughArguments = new ArrayList<String>();
 		FakeCucumberBackendFactory cucumberBackendFactory = new FakeCucumberBackendFactory();
 		RuntimeConfiguration runtimeConfiguration = new RuntimeConfiguration(0, cucumberPassthroughArguments, null, featurePaths,
-				null, false, null, false, null, false);
+				null, false, null, false, null, false, false);
 		CucumberRuntimeFactory runtimeFactory = new CucumberRuntimeFactory(runtimeConfiguration, cucumberBackendFactory, Thread
 				.currentThread().getContextClassLoader());
 		runtimeFactory.getRuntime(cucumberArguments);
@@ -58,7 +58,7 @@ public class CucumberRuntimeFactoryTest {
 		List<String> cucumberPassthroughArguments = new ArrayList<String>();
 		FakeCucumberBackendFactory cucumberBackendFactory = new FakeCucumberBackendFactory();
 		RuntimeConfiguration runtimeConfiguration = new RuntimeConfiguration(0, cucumberPassthroughArguments, null, featurePaths,
-				null, false, null, false, null, false);
+				null, false, null, false, null, false, false);
 		CucumberRuntimeFactory runtimeFactory = new CucumberRuntimeFactory(runtimeConfiguration, cucumberBackendFactory, Thread
 				.currentThread().getContextClassLoader());
 		assertThat(runtimeFactory.getRuntime(cucumberArguments)).isInstanceOf(Runtime.class);
@@ -70,7 +70,7 @@ public class CucumberRuntimeFactoryTest {
 		List<String> cucumberPassthroughArguments = new ArrayList<String>();
 		FakeCucumberBackendFactory cucumberBackendFactory = new FakeCucumberBackendFactory();
 		RuntimeConfiguration runtimeConfiguration = new RuntimeConfiguration(0, cucumberPassthroughArguments, null, featurePaths,
-				null, false, null, false, null, true);
+				null, false, null, false, null, true, false);
 		CucumberRuntimeFactory runtimeFactory = new CucumberRuntimeFactory(runtimeConfiguration, cucumberBackendFactory, Thread
 				.currentThread().getContextClassLoader());
 		assertThat(runtimeFactory.getRuntime(cucumberArguments)).isInstanceOf(ThreadLoggedRuntime.class);