+++ src/test/java/com/bishnet/cucumber/parallel/runtime/RerunFileBuilderTest.java
@@ -93,6 +93,6 @@ public class RerunFileBuilderTest {
 	}
 
 	private RuntimeConfiguration getRuntimeConfiguration(List<String> featureParsingArguments) {
-		return new RuntimeConfiguration(0, null, featureParsingArguments, null, null, false, null, false, null, false);
+		return new RuntimeConfiguration(0, null, featureParsingArguments, null, null, false, null, false, null, false, false);
 	}
 }